starterkit: use map table instead of version settings
Each version of Android/Yocto require different file name,
renesas-bsp-rom-writer has its setting files to know which files are
needed. But, if no big change was added, it will be very duplicated
files. We want to use symbolic link for it but we can't,
because renesas-bsp-rom-writer is used from both Windows/Linux.
To solve this issue, this patch uses map table.

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>