starterkit: windows: indicate manual stop
We are not sure why, but TeraTerm doesn't send 0x01 if it was

	1) You selected "use Mot file" on TeraTerm
	2) U-boot has Auto Run setting (= bootcmd)
	3) Power ON StarterKit, and wait to start Auto Run
	4) Stop Auto Run manually using Enter and/or Ctrl-C
	   multiple times.

The macro try to stop Auto Run by using Ctrl-C,
and send CPLD settings after that.

	;; Power ON message here

	;; Ctrl-C loop here

	sendln "cpld write 0x00 0x802181fe"
	sendln "cpld write 0x80 0x01"
	wait "plase send !"

But TeraTerm might not send last 0x01.

	=> cpld write 0x00 0x802181fe
	=> cpld write 0x80
	                    ^^^^

It seems TeraTerm has this kind of bug (can't send 0x01) before,
but it was already solved by v4.75. We don't why this happen.

This patch indicate manually stop the Auto Run,
and add FAQ for it.

Reported-by: Yusuke Goda <yusuke.goda.sx@renesas.com>
Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>