Add new script
Current ROM writer is unfortunately not so user-friendly,
thus, it looks very complex for beginner user.
The issue is that beginner is easy to confuse/misunderstand
by many reasons.

1st reason is that ROM writer command requests select option
for Yocto or for Android which is used to know the nessesary
ROM file and its load postion

In Yocto case, using ROM writer is everything, but in Android case,
it needs 2 steps. 1st is writing ROM, and 2nd is Android settings.

But beginner will misunderstand that ROM writer command will setup
all Android settings because he selected Android via option though
it needs other Android setting command.
Of couse these are explained in README, but beginner doesn't read it.

2nd reason is that, many limitation/note/prepare are needed and these
are indicated at README. But it has many READMEs. for board, for linux,
for windows, etc. And same as 1st reason, beginner doesn't read it..

Becase ROM writing needs many steps and many limitation/note,
interrctive command is better idea from beginner and/or user-friendly
point of view.

This patch adds new style of script for it.

The new script will indicate nesessary message at nesessary timing,
thus user doesn't need to check README.
And also it is easy to adjust to new style board (Current script is
deeply depends on StarterKit/Salvator style).

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>