Change the treatment of mot
Current ROM writer is assuming that we can compile mot file
by ourself. But it seems we can't for some SoC.
And it is also assuming that the mot file is always same
for same SoC, but it seems we need to switch for some SoC.
This patch doesn't assume the mot file on base.py, and each
board script handles it.

This is prepare for S4 beta3 support

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>